concierge_packages/launcher/pyproject.toml
[project]
name = "launch-concierge"
version = "0.4.0"
requires-python = ">=3.10"
dependencies = [
    "script-builder",
    "isi-util",
    "concierge-util",
    "packaging",
    "python-dotenv"
]

[project.scripts]
install_concierge = "launch_concierge.install:install"
relaunch_concierge = "launch_concierge.relaunch:relaunch"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"